requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
lxml>=5.0.0
tqdm>=4.66.0
